requirements.txt
gunicorn==20.1.0
python-dotenv==0.20.0
Flask==2.1.2
flask-restful==0.3.9
flask-mongoengine==1.0.0
flask-jwt-extended==4.4.0
flask-cors==3.0.10
apscheduler==3.9.1
pymongo==4.1.1
mongoengine==0.24.1
firebase_admin==5.2.0
pandas==1.4.2
numpy==1.22.3
matplotlib==3.5.1
hurry.filesize==0.9